requirements.txt
atlassian-python-api
termcolor
coloredlogs
markdown2
